폴더구조 변경: domain, services
