+++ src/ticketing/application/domain/repositories/iconcert.repository.ts
@@ -1,0 +1,9 @@
+import { Concert } from '../models/concert';
+import { ConcertSchedule } from '../models/concert-schedule';
+import { Seat } from '../models/seat';
+
+export interface IConcertRepository {
+	findConcerts(): Promise<Concert[]>;
+	findSchedules(concertId: number): Promise<ConcertSchedule[]>;
+	findSeats(scheduleId: number): Promise<Seat[]>;
+}
