+++ src/ticketing/application/domain/repositories/ireservation.repository.ts
@@ -1,0 +1,5 @@
+import { Reservation, ReservationProps } from '../models/reservation';
+
+export interface IReservationRepository {
+	create(props: ReservationProps): Promise<Reservation>;
+}
