+++ src/ticketing/application/domain/repositories/iseat.repository.ts
@@ -1,0 +1,6 @@
+import { Seat, SeatStatus } from '../models/seat';
+
+export interface ISeatRepository {
+	findSeatById(seatId: number): Promise<Seat>;
+	updateStatus(seatId: number, status: SeatStatus): Promise<boolean>;
+}
